fix: Have a root component
